Updated Mg.Metal to 5.0.4
