+++ Magnesium.Metal/Driver/AmtPresentationSurface.cs
@@ -1,8 +1,16 @@
 ﻿using System;
+using MetalKit;
+
 namespace Magnesium.Metal
 {
 	public class AmtPresentationSurface : IMgPresentationSurface
 	{
+		readonly MTKView mApplicationView;
+		public AmtPresentationSurface(MTKView view)
+		{
+			mApplicationView = view;
+		}
+
 		public IMgSurfaceKHR Surface
 		{
 			get
@@ -16,9 +24,9 @@ namespace Magnesium.Metal
 			
 		}
 
-		public void Initialize()
+		public void Initialize(uint width, uint height)
 		{
-			
+			mApplicationView.DrawableSize = new CoreGraphics.CGSize((nfloat)width, (nfloat)height);
 		}
 	}
 }
