+++ Magnesium.Metal/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Magnesium" version="4.0.0" targetFramework="net45" />
+  <package id="Magnesium" version="5.0.4" targetFramework="net45" />
 </packages>
