+++ examples/MetalSample/MetalSample/MetalSample.csproj
@@ -53,13 +53,10 @@
     <Reference Include="System.IO.Compression" />
     <Reference Include="System.Net.Http" />
     <Reference Include="Magnesium">
-      <HintPath>..\packages\Magnesium.4.0.0\lib\portable-net45+win+wpa81+MonoAndroid10+xamarinios10+MonoTouch10\Magnesium.dll</HintPath>
-    </Reference>
-    <Reference Include="Magnesium.Metal">
-      <HintPath>..\packages\Magnesium.Metal.4.0.1\lib\portable-net45+netcore45+wpa81\Magnesium.Metal.dll</HintPath>
+      <HintPath>..\packages\Magnesium.5.0.4\lib\portable45-net45+win8+wpa81\Magnesium.dll</HintPath>
     </Reference>
     <Reference Include="SimpleInjector">
-      <HintPath>..\packages\SimpleInjector.3.2.5\lib\net45\SimpleInjector.dll</HintPath>
+      <HintPath>..\packages\SimpleInjector.3.2.7\lib\net45\SimpleInjector.dll</HintPath>
       <Private>False</Private>
     </Reference>
   </ItemGroup>
@@ -92,7 +89,7 @@
     <Compile Include="GameViewController.designer.cs">
       <DependentUpon>GameViewController.cs</DependentUpon>
     </Compile>
-    <Compile Include="MgGraphicsConfiguration.cs" />
+    <Compile Include="SpinningCube.cs" />
   </ItemGroup>
   <ItemGroup>
     <InterfaceDefinition Include="MainMenu.xib" />
@@ -107,6 +104,12 @@
     <BundleResource Include="Resources\Fragment.metal">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </BundleResource>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\Magnesium.Metal\Magnesium.Metal.csproj">
+      <Project>{32B6046B-4432-4DDF-9718-EB5C3A1A0285}</Project>
+      <Name>Magnesium.Metal</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Mac\Xamarin.Mac.CSharp.targets" />
 </Project>
