+++ examples/MetalSample/MetalSample/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Magnesium" version="4.0.0" targetFramework="net45" />
-  <package id="Magnesium.Metal" version="4.0.1" targetFramework="net45" />
-  <package id="SimpleInjector" version="3.2.5" targetFramework="net45" />
+  <package id="Magnesium" version="5.0.4" targetFramework="net45" />
+  <package id="SimpleInjector" version="3.2.7" targetFramework="net45" />
 </packages>
